Extra peer address more reliably
Using transport.peer doesn't work that great when going through a proxy. Best thing to do is to pull in the following URL:

- x-real-ip
- x-forwarded-ip
- peer

That order should more reliabily allow us to determine what the originating connection IP for the client might be. Useful for the audit logs and registration connection information.