refs #10140 Don't throw when uuid doesn't resolve
The code was throwing an exception of `NoneType` not having `record_type_` attribute. This was since in certain cases the uuid provided to `db.get` may not resolve to a proper object. Sometimes this happends due to aggressive culling or other reasons. In this case `get` should just return `None`.