IMP handle out of band authenticate requests
In weird cases we see messages like:

```
2022-08-03T08:17:38-0700 [Router         20] Traceback (most recent call last):
  File "/usr/local/lib/pypy3.8/dist-packages/autobahn/wamp/websocket.py", line 103, in onMessage
    self._session.onMessage(msg)
  File "/app/crossbar/router/session.py", line 658, in onMessage
    self._router.process(self, msg)
  File "/app/crossbar/router/router.py", line 520, in process
    super().process( session, msg )
  File "/app/crossbar/router/router.py", line 330, in process
    raise ProtocolError("Unexpected message {0}".format(msg.__class__))
autobahn.wamp.exception.ProtocolError: Unexpected message <class 'autobahn.wamp.message.Authenticate'>
```

This is troublesome since this has a habit of throwing the client into a disconnect/reconnect loop. With this modification, we'll just throw the already authenticated information back to the user.

There might be room for improvement where we validate the user's credentials again (and verify that we're not trying to do something weird like switch authids) but we're not going to worry about it for now.